requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
wandb>=0.15.0

# HuggingFace
transformers>=4.30.0
datasets>=2.14.0
tokenizers>=0.13.0

# Model-specific (local development only)
einops>=0.6.0

# For GPT-2 training
# nanoGPT or minGPT will be cloned separately

# Additional utilities
pyyaml>=6.0
tensorboard>=2.13.0
